Add link/url_for and link_to helpers
